slette bruker og logge ut

+++ Data/tblUser/deleteAll.sql
@@ -1,2 +1,3 @@
-DELETE FROM "tblUser"
-WHERE "fdUserID" = $1;
+DELETE FROM "tblListItem" WHERE "fdUserID" = $1;
+DELETE FROM "tblToDoList" WHERE "fdUserID" = $1;
+DELETE FROM "tblUser" WHERE "fdUserID" = $1;
